+++ docs/sphinx/changelog.rst
@@ -6,7 +6,27 @@ Changelog
 #########
 
 ******************
-0.1.0 (unreleased)
+0.2.1 (unreleased)
+******************
+
+******************
+0.2.0 (06-12-2025)
+******************
+
+New Features
+============
+- Added a dof velocity gradient driven deformation material model DefinedPlasticEvolution (:pull:`7`). By `Nathan Miller`_.
+
+Breaking Changes
+================
+- Changed BasicReactingSolid to BasicSolid since it is more accurate (:pull:`7`). By `Nathan Miller`_.
+
+Release
+=======
+- Release version (:pull:`8`). By `Nathan Miller`_.
+
+******************
+0.1.0 (05-21-2025)
 ******************
 
 New Features
@@ -18,6 +38,9 @@ New Features
 
 Internal Changes
 ================
+- Prepared the library for conda packaging (:pull:`5`). By `Nathan Miller`_.
 
-Breaking Changes
-================
+
+Release
+=======
+- Release version (:pull:`6`). By `Nathan Miller`_.